Use DHCP on the management network
Configure & use a DHCP server on the wbench VM to configure the managment ports
on the switches, just as the real workbench does it.
Use the correct IPs in the hosts file for lead1 & leaf1.
Use fixed MACs for eth1 on both leaf1 & leaf2 so that the DHCP server can match
them.