Enable remapping of management interfaces
Add an init script which can swap the eth0 & eth1 interfaces. Vagrant always
requires the first interface (eth0) to be a NAT device, but if this init
script is run (& enabled) at the end of the Vagrant provisioning process it can
"cheat" Vagrant and allow us to have the "internal" management interface as
eth0, with the Vagrant NAT interface on eth1.